fix: remaining dev-review comments
